Move addons and localhelm under workloads
