+++ pkg/workloads/addons/ensure.go
