+++ pkg/workloads/addons/helpers.go
