+++ pkg/workloads/addons/manifest.go
