+++ pkg/workloads/addons/manifest_test.go
